fixed img css styles. moved all pages to components to set up routings later
